Reference the example file doc/my_test_suite.lua in the documentation.
